refactor misplaced class
